Upload project
Uploaded my old finished project for archiving. This commit includes the following features. Added forms for adding stations and measurements as well as actions for deleting them. Users can now register, log in and log out. Weather stations are now assigned to users. Dashboard and detail page now show the current weather with appropriate icons and trends. Weather stations are now displayed on a map in the dashboard. The detail page of a station now has a function to read data automatically via OpenWeatherMap.